requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.1
django-crispy-forms==1.9.2
django-geojson==2.12.0
django-heroku==0.3.1
django-leaflet==0.24.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
html5lib==1.1
idna==2.8
importlib-metadata==1.7.0
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-server==1.2.0
kiwisolver==1.2.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
more-itertools==8.4.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.19.2
openpyxl==3.0.5
packaging==20.4
paho-mqtt==1.4.0
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.11.4
pluggy==0.13.1
pmdarima==1.7.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
python-dateutil==2.8.1
pytz==2019.1
pyzmq==19.0.2
requests==2.22.0
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.3.0
statsmodels==0.11.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
timer==0.1.0
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.3
virtualenv==20.1.0
virtualenv-clone==0.5.4
waitress==1.4.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
xlrd==1.2.0
zipp==3.1.0
